Problem: allowed recursive functions
Now I have no problem with recursion, except that it's a mystery
to anyone looking at code, and inherently unstable, and useless
except as a mental ego booster.

Solution: only update dictionary head when committing a function
so that references to the same function won't work. As there are
no forward references, recursion is now impossible.

Also, simplifications.