Problem: core loop is no longer correct
Solution: rewrite as a REPL (read, evaluate, print, loop) engine.
In fact this doesn't do the REPL, that would be the main zs tool.
However I like the name and the culture it comes with.

The state machine does not yet handle function definitions.
However it's generating and running bytecode nicely already.