Problem: "K: (1000 *), K (1 2)" gives "1000 2" instead of "1000 2000"
Caused by error in zs_pipe_array_pull which was pulling operand to
end of input pipe instead of start. When input was empty this worked
and when input was not empty (as after unnest), this failed.

Solution: pull operand to start of input, not end.

Also added test case for this.